homework-7 done
