+++ js/generateInteger.js
@@ -1,0 +1,7 @@
+export const generateInteger = (min, max) => {
+    min = Math.ceil(min);
+    max = Math.floor(max);
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export default generateInteger;
